// src/api-spec/protobuf/gen/js/tdex/v2/types_pb.d.ts
export interface UnblindedInput {
    /**
     * Index of the transaction input of reference.
     *
     * @generated from protobuf field: uint32 index = 1;
     */
    index: number;
    /**
     * Unblinded asset.
     *
     * @generated from protobuf field: string asset = 2;
     */
    asset: string;
    /**
     * Unblinded amount.
     *
     * @generated from protobuf field: uint64 amount = 3;
     */
    amount: string;
    /**
     * Asset blinider when blinding the prevout.
     *
     * @generated from protobuf field: string asset_blinder = 4;
     */
    assetBlinder: string;
    /**
     * Amount blinider used when blinding the prevout.
     *
     * @generated from protobuf field: string amount_blinder = 5;
     */
    amountBlinder: string;
}
// ...
declare class UnblindedInput$Type extends MessageType<UnblindedInput> {
    constructor();
    create(value?: PartialMessage<UnblindedInput>): UnblindedInput;
    internalBinaryRead(reader: IBinaryReader, length: number, options: BinaryReadOptions, target?: UnblindedInput): UnblindedInput;
    internalBinaryWrite(message: UnblindedInput, writer: IBinaryWriter, options: BinaryWriteOptions): IBinaryWriter;
}
/**
 * @generated MessageType for protobuf message tdex.v2.UnblindedInput
 */
export declare const UnblindedInput: UnblindedInput$Type;